# docs/api/radiance-openapi.json
{
  "components": {
    "schemas": {
      "AssemblyAssetFallbackResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_key": {
            "title": "Asset Key",
            "type": "string"
          },
          "fallback_asset_key": {
            "title": "Fallback Asset Key",
            "type": "string"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "title": "Instance Ids",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "asset_key",
          "fallback_asset_key",
          "instance_ids",
          "reason"
        ],
        "title": "AssemblyAssetFallbackResponse",
        "type": "object"
      },
      "AssemblyAssetsResponse": {
        "additionalProperties": false,
        "properties": {
          "anchors": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchors"
          },
          "manifest": {
            "title": "Manifest",
            "type": "string"
          }
        },
        "required": [
          "manifest"
        ],
        "title": "AssemblyAssetsResponse",
        "type": "object"
      },
      "AssemblyAxisMappingResponse": {
        "additionalProperties": false,
        "properties": {
          "cad_horizontal": {
            "items": {
              "enum": [
                "x",
                "z"
              ],
              "type": "string"
            },
            "title": "Cad Horizontal",
            "type": "array"
          },
          "cad_vertical": {
            "const": "y",
            "title": "Cad Vertical",
            "type": "string"
          },
          "layout_horizontal": {
            "items": {
              "enum": [
                "x",
                "y"
              ],
              "type": "string"
            },
            "title": "Layout Horizontal",
            "type": "array"
          },
          "layout_vertical": {
            "const": "z",
            "title": "Layout Vertical",
            "type": "string"
          }
        },
        "required": [
          "cad_horizontal",
          "cad_vertical",
          "layout_horizontal",
          "layout_vertical"
        ],
        "title": "AssemblyAxisMappingResponse",
        "type": "object"
      },
      "AssemblyFixturePointResponse": {
        "additionalProperties": false,
        "properties": {
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          },
          "z": {
            "title": "Z",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "z"
        ],
        "title": "AssemblyFixturePointResponse",
        "type": "object"
      },
      "AssemblyInstanceResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_fallback_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Fallback Key"
          },
          "asset_key": {
            "title": "Asset Key",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "layout_source": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Source"
          },
          "layout_type": {
            "title": "Layout Type",
            "type": "string"
          },
          "module_count": {
            "title": "Module Count",
            "type": "integer"
          },
          "orient": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orient"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/AssemblyFixturePointResponse"
            },
            "title": "Points",
            "type": "array"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssemblyFixturePointResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "shape": {
            "enum": [
              "linear",
              "corner",
              "centerpiece",
              "fixture",
              "unknown"
            ],
            "title": "Shape",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "yaw_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yaw Deg"
          }
        },
        "required": [
          "id",
          "layout_type",
          "orient",
          "module_count",
          "shape",
          "asset_key",
          "asset_fallback_key",
          "points",
          "warnings"
        ],
        "title": "AssemblyInstanceResponse",
        "type": "object"
      },
      "AssemblyModuleAssetResponse": {
        "additionalProperties": false,
        "properties": {
          "anchor_source": {
            "anyOf": [
              {
                "const": "module_nodes",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Source"
          },
          "fallback_asset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Asset Key"
          },
          "high": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "medium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medium"
          },
          "optional": {
            "default": false,
            "title": "Optional",
            "type": "boolean"
          },
          "proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy"
          }
        },
        "required": [
          "high",
          "medium",
          "proxy"
        ],
        "title": "AssemblyModuleAssetResponse",
        "type": "object"
      },
...
      "AssemblyRoomResponse": {
        "additionalProperties": false,
        "properties": {
          "length_ft": {
            "title": "Length Ft",
            "type": "number"
          },
          "length_m": {
            "title": "Length M",
            "type": "number"
          },
          "mount_z_m": {
            "title": "Mount Z M",
            "type": "number"
          },
          "width_ft": {
            "title": "Width Ft",
            "type": "number"
          },
          "width_m": {
            "title": "Width M",
            "type": "number"
          }
        },
        "required": [
          "length_ft",
          "width_ft",
          "length_m",
          "width_m",
          "mount_z_m"
        ],
        "title": "AssemblyRoomResponse",
        "type": "object"
      },
      "AssemblySceneResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_fallbacks_used": {
            "items": {
              "$ref": "#/components/schemas/AssemblyAssetFallbackResponse"
            },
            "title": "Asset Fallbacks Used",
            "type": "array"
          },
          "assets": {
            "$ref": "#/components/schemas/AssemblyAssetsResponse"
          },
          "axis_mapping": {
            "$ref": "#/components/schemas/AssemblyAxisMappingResponse"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "fixture_counts_by_asset_key": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Fixture Counts By Asset Key",
            "type": "object"
          },
          "fixture_counts_by_layout_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Fixture Counts By Layout Type",
            "type": "object"
          },
          "fspm_metrics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fspm Metrics"
          },
          "instances": {
            "items": {
              "$ref": "#/components/schemas/AssemblyInstanceResponse"
            },
            "title": "Instances",
            "type": "array"
          },
          "missing_asset_keys": {
            "items": {
              "type": "string"
            },
            "title": "Missing Asset Keys",
            "type": "array"
          },
          "mode": {
            "enum": [
              "SMD",
              "Competitor",
              "1000W HPS"
            ],
            "title": "Mode",
            "type": "string"
          },
          "mode_label": {
            "title": "Mode Label",
            "type": "string"
          },
          "module_assets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AssemblyModuleAssetResponse"
            },
            "title": "Module Assets",
            "type": "object"
          },
          "placement_strategy": {
            "default": "anchor_fit",
            "enum": [
              "anchor_fit",
              "single_fixture_center"
            ],
            "title": "Placement Strategy",
            "type": "string"
          },
          "plants": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plants"
          },
          "room": {
            "$ref": "#/components/schemas/AssemblyRoomResponse"
          },
          "schema_version": {
            "const": 3,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_units": {
            "const": "millimeters",
            "title": "Source Units",
            "type": "string"
          },
          "system": {
            "enum": [
              "proposed_led_system",
              "conventional_led_system",
              "hps_1000w_system"
            ],
            "title": "System",
            "type": "string"
          },
          "units": {
            "const": "meters",
            "title": "Units",
            "type": "string"
          },
          "viewer_scale": {
            "title": "Viewer Scale",
            "type": "number"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "system",
          "mode",
          "mode_label",
          "display_name",
          "units",
          "source_units",
          "viewer_scale",
          "axis_mapping",
          "room",
          "assets",
          "module_assets",
          "instances",
          "fixture_counts_by_layout_type",
          "fixture_counts_by_asset_key",
          "missing_asset_keys",
          "asset_fallbacks_used",
          "warnings"
        ],
        "title": "AssemblySceneResponse",
        "type": "object"
      },
...
      "PublicErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "const": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error",
          "message",
          "correlation_id"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Horticulture Lighting Simulator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/radiance/assembly-scene": {
      "get": {
        "operationId": "get_radiance_assembly_scene",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "SMD",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "execution_mode",
            "required": false,
            "schema": {
              "default": "precomputed",
              "title": "Execution Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sim_mode",
            "required": false,
            "schema": {
              "default": "standard",
              "title": "Sim Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_ppfd",
            "required": false,
            "schema": {
              "default": 1000.0,
              "title": "Target Ppfd",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "peak_capping_enabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Peak Capping Enabled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "match_system_ppe",
            "required": false,
            "schema": {
              "default": false,
              "title": "Match System Ppe",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "length_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Length Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "width_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Width Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "w_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "W Min"
            }
          },
          {
            "in": "query",
            "name": "w_max",
            "required": false,
            "schema": {
              "default": 100.0,
              "title": "W Max",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "competitor_layout",
            "required": false,
            "schema": {
              "default": "full",
              "title": "Competitor Layout",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hps_coverage_ft",
            "required": false,
            "schema": {
              "default": 4.0,
              "title": "Hps Coverage Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_ies_variant",
            "required": false,
            "schema": {
              "default": "karma",
              "title": "Hps Ies Variant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mount_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Mount Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "sp_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Sp Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Hps Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "basis_backend",
            "required": false,
            "schema": {
              "default": "rtrace",
              "title": "Basis Backend",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plants_enabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Plants Enabled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "plant_seed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Seed"
            }
          },
          {
            "in": "query",
            "name": "plant_rows",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Rows"
            }
          },
          {
            "in": "query",
            "name": "plant_columns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Columns"
            }
          },
          {
            "in": "query",
            "name": "plant_spacing_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Spacing M"
            }
          },
          {
            "in": "query",
            "name": "plant_height_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Height M"
            }
          },
          {
            "in": "query",
            "name": "plant_canopy_radius_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Canopy Radius M"
            }
          },
          {
            "in": "query",
            "name": "plant_leaf_count",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Leaf Count"
            }
          },
          {
            "in": "query",
            "name": "plant_growth_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Growth Stage"
            }
          },
          {
            "in": "query",
            "name": "fspm_target_ppfd_umol_m2_s",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fspm Target Ppfd Umol M2 S"
            }
          },
          {
            "in": "query",
            "name": "fspm_target_tolerance_umol_m2_s",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fspm Target Tolerance Umol M2 S"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "artifact_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssemblySceneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Radiance Assembly Scene"
      }
    },
...
  }
}